// src/shared/types/obsActions.ts
export interface ObsActionBase {
  type: string;
}
// ...
export interface CreateInputAction extends ObsActionBase {
  type: 'createInput';
  inputName: string;
  inputKind: string;
  inputSettings?: Record<string, unknown>;
  sceneName?: string;
  sceneItemEnabled?: boolean;
}

export interface SetInputSettingsAction extends ObsActionBase {
  type: 'setInputSettings';
  inputName: string;
  inputSettings: Record<string, unknown>;
  overlay?: boolean;
}

export interface SetSceneItemEnabledAction extends ObsActionBase {
  type: 'setSceneItemEnabled';
  sceneName: string;
  sourceName: string;
  sceneItemEnabled: boolean;
}

export interface GetInputSettingsAction extends ObsActionBase {
  type: 'getInputSettings';
  inputName: string;
}

export interface GetSceneItemListAction extends ObsActionBase {
  type: 'getSceneItemList';
  sceneName: string;
}

export interface SetCurrentProgramSceneAction extends ObsActionBase {
  type: 'setCurrentProgramScene';
  sceneName: string;
}

export interface SetVideoSettingsAction extends ObsActionBase {
  type: 'setVideoSettings';
  videoSettings: OBSVideoSettings;
}

export interface CreateSceneAction extends ObsActionBase {
  type: 'createScene';
  sceneName: string;
}

export interface RemoveInputAction extends ObsActionBase {
  type: 'removeInput';
  inputName: string;
}

export interface SetSceneItemTransformAction extends ObsActionBase {
  type: 'setSceneItemTransform';
  sceneName: string;
  sourceName: string;
  transform: {
    positionX?: number;
    positionY?: number;
    scaleX?: number;
    scaleY?: number;
    rotation?: number;
    alignment?: number;
  };
}

export interface CreateSourceFilterAction extends ObsActionBase {
  type: 'createSourceFilter';
  sourceName: string;
  filterName: string;
  filterKind: string;
  filterSettings?: Record<string, unknown>;
}

export interface SetInputVolumeAction extends ObsActionBase {
  type: 'setInputVolume';
  inputName: string;
  inputVolumeMul?: number;
  inputVolumeDb?: number;
}

export interface SetInputMuteAction extends ObsActionBase {
  type: 'setInputMute';
  inputName: string;
  inputMuted: boolean;
}

export interface StartVirtualCamAction extends ObsActionBase {
  type: 'startVirtualCam';
}

export interface StopVirtualCamAction extends ObsActionBase {
  type: 'stopVirtualCam';
}

export interface SaveScreenshotAction extends ObsActionBase {
  type: 'saveScreenshot';
  imageFormat: string;
  imageFilePath: string;
  imageWidth?: number;
  imageHeight?: number;
}

export interface StartReplayBufferAction extends ObsActionBase {
  type: 'startReplayBuffer';
}

export interface SaveReplayBufferAction extends ObsActionBase {
  type: 'saveReplayBuffer';
}

export interface SetSourceFilterIndexAction extends ObsActionBase {
  type: 'setSourceFilterIndex';
  sourceName: string;
  filterName: string;
  filterIndex: number;
}

export interface SetSourceFilterNameAction extends ObsActionBase {
  type: 'setSourceFilterName';
  sourceName: string;
  filterName: string;
  newFilterName: string;
}

export interface DuplicateSourceFilterAction extends ObsActionBase {
  type: 'duplicateSourceFilter';
  sourceName: string;
  filterName: string;
  newFilterName: string;
}

export interface TriggerStudioModeTransitionAction extends ObsActionBase {
  type: 'triggerStudioModeTransition';
}

export interface SetInputAudioMonitorTypeAction extends ObsActionBase {
  type: 'setInputAudioMonitorType';
  inputName: string;
  monitorType:
    | 'OBS_MONITORING_TYPE_NONE'
    | 'OBS_MONITORING_TYPE_MONITOR_ONLY'
    | 'OBS_MONITORING_TYPE_MONITOR_AND_OUTPUT';
}

export interface SetSceneItemBlendModeAction extends ObsActionBase {
  type: 'setSceneItemBlendMode';
  sceneName: string;
  sourceName: string;
  blendMode: string;
}

export interface SetSceneNameAction extends ObsActionBase {
  type: 'setSceneName';
  sceneName: string;
  newSceneName: string;
}

export interface RefreshBrowserSourceAction extends ObsActionBase {
  type: 'refreshBrowserSource';
  inputName: string;
}

export interface GetLogFileListAction extends ObsActionBase {
  type: 'getLogFileList';
}

export interface GetLogFileAction extends ObsActionBase {
  type: 'getLogFile';
  logFile: string;
}

export interface SetStudioModeEnabledAction extends ObsActionBase {
  type: 'setStudioModeEnabled';
  enabled: boolean;
}

export interface ToggleStudioModeAction extends ObsActionBase {
  type: 'toggleStudioMode';
}

export interface TriggerHotkeyByNameAction extends ObsActionBase {
  type: 'triggerHotkeyByName';
  hotkeyName: string;
}

export interface TriggerHotkeyByKeySequenceAction extends ObsActionBase {
  type: 'triggerHotkeyByKeySequence';
  keyId: string;
  keyModifiers: { shift: boolean; control: boolean; alt: boolean; command: boolean };
}

export interface ToggleStreamAction extends ObsActionBase {
  type: 'toggleStream';
}

export interface ToggleRecordAction extends ObsActionBase {
  type: 'toggleRecord';
}

export interface GetSourceFilterListAction extends ObsActionBase {
  type: 'getSourceFilterList';
  sourceName: string;
}

export interface GetSourceFilterDefaultSettingsAction extends ObsActionBase {
  type: 'getSourceFilterDefaultSettings';
  filterKind: string;
}

export interface GetSourceFilterSettingsAction extends ObsActionBase {
  type: 'getSourceFilterSettings';
  sourceName: string;
  filterName: string;
}

export interface SetSourceFilterSettingsAction extends ObsActionBase {
  type: 'setSourceFilterSettings';
  sourceName: string;
  filterName: string;
  filterSettings: Record<string, unknown>;
  overlay?: boolean;
}

export interface SetSourceFilterEnabledAction extends ObsActionBase {
  type: 'setSourceFilterEnabled';
  sourceName: string;
  filterName: string;
  filterEnabled: boolean;
}

export interface RemoveSourceFilterAction extends ObsActionBase {
  type: 'removeSourceFilter';
  sourceName: string;
  filterName: string;
}

export interface GetInputDefaultSettingsAction extends ObsActionBase {
  type: 'getInputDefaultSettings';
  inputKind: string;
}

export interface GetOutputListAction extends ObsActionBase {
  type: 'getOutputList';
}

export interface GetOutputStatusAction extends ObsActionBase {
  type: 'getOutputStatus';
  outputName: string;
}

export interface StartOutputAction extends ObsActionBase {
  type: 'startOutput';
  outputName: string;
}

export interface StopOutputAction extends ObsActionBase {
  type: 'stopOutput';
  outputName: string;
}

export interface GetOutputSettingsAction extends ObsActionBase {
  type: 'getOutputSettings';
  outputName: string;
}

export interface SetOutputSettingsAction extends ObsActionBase {
  type: 'setOutputSettings';
  outputName: string;
  outputSettings: Record<string, unknown>;
}

export interface GetSceneTransitionListAction extends ObsActionBase {
  type: 'getSceneTransitionList';
}

export interface GetCurrentSceneTransitionAction extends ObsActionBase {
  type: 'getCurrentSceneTransition';
}

export interface SetCurrentSceneTransitionAction extends ObsActionBase {
  type: 'setCurrentSceneTransition';
  transitionName: string;
}

export interface SetSceneTransitionDurationAction extends ObsActionBase {
  type: 'setSceneTransitionDuration';
  transitionDuration: number;
}

export interface GetSceneTransitionCursorAction extends ObsActionBase {
  type: 'getSceneTransitionCursor';
}

export interface GetMediaInputStatusAction extends ObsActionBase {
  type: 'getMediaInputStatus';
  inputName: string;
}

export interface SetMediaInputCursorAction extends ObsActionBase {
  type: 'setMediaInputCursor';
  inputName: string;
  mediaCursor: number;
}

export interface OffsetMediaInputCursorAction extends ObsActionBase {
  type: 'offsetMediaInputCursor';
  inputName: string;
  mediaCursorOffset: number;
}

export interface TriggerMediaInputActionAction extends ObsActionBase {
  type: 'triggerMediaInputAction';
  inputName: string;
  mediaAction: string;
}

export interface GetCurrentPreviewSceneAction extends ObsActionBase {
  type: 'getCurrentPreviewScene';
}

export interface SetCurrentPreviewSceneAction extends ObsActionBase {
  type: 'setCurrentPreviewScene';
  sceneName: string;
}

export interface GetSceneItemLockedAction extends ObsActionBase {
  type: 'getSceneItemLocked';
  sceneName: string;
  sceneItemId: number;
}

export interface SetSceneItemLockedAction extends ObsActionBase {
  type: 'setSceneItemLocked';
  sceneName: string;
  sceneItemId: number;
  sceneItemLocked: boolean;
}

export interface GetSceneItemIndexAction extends ObsActionBase {
  type: 'getSceneItemIndex';
  sceneName: string;
  sceneItemId: number;
}

export interface SetSceneItemIndexAction extends ObsActionBase {
  type: 'setSceneItemIndex';
  sceneName: string;
  sceneItemId: number;
  sceneItemIndex: number;
}

export interface CreateSceneItemAction extends ObsActionBase {
  type: 'createSceneItem';
  sceneName: string;
  sourceName: string;
  sceneItemEnabled: boolean;
}

export interface RemoveSceneItemAction extends ObsActionBase {
  type: 'removeSceneItem';
  sceneName: string;
  sceneItemId: number;
}

export interface GetStatsAction extends ObsActionBase {
  type: 'getStats';
}

export interface GetVersionAction extends ObsActionBase {
  type: 'getVersion';
}
// ...
export type ObsAction =
  | CreateInputAction
  | SetInputSettingsAction
  | SetSceneItemEnabledAction
  | GetInputSettingsAction
  | GetSceneItemListAction
  | SetCurrentProgramSceneAction
  | SetVideoSettingsAction
  | CreateSceneAction
  | RemoveInputAction
  | SetSceneItemTransformAction
  | CreateSourceFilterAction
  | SetInputVolumeAction
  | SetInputMuteAction
  | StartVirtualCamAction
  | StopVirtualCamAction
  | SaveScreenshotAction
  | StartReplayBufferAction
  | SaveReplayBufferAction
  | SetSourceFilterIndexAction
  | SetSourceFilterNameAction
  | DuplicateSourceFilterAction
  | TriggerStudioModeTransitionAction
  | SetInputAudioMonitorTypeAction
  | SetSceneItemBlendModeAction
  | RefreshBrowserSourceAction
  | GetLogFileListAction
  | GetLogFileAction
  | ToggleStudioModeAction
  | SetStudioModeEnabledAction
  | TriggerHotkeyByNameAction
  | TriggerHotkeyByKeySequenceAction
  | GetSourceFilterListAction
  | GetSourceFilterDefaultSettingsAction
  | GetSourceFilterSettingsAction
  | SetSourceFilterSettingsAction
  | SetSourceFilterEnabledAction
  | RemoveSourceFilterAction
  | ToggleStreamAction
  | ToggleRecordAction
  | GetInputDefaultSettingsAction
  | GetOutputListAction
  | GetOutputStatusAction
  | StartOutputAction
  | StopOutputAction
  | GetOutputSettingsAction
  | SetOutputSettingsAction
  | GetSceneTransitionListAction
  | GetCurrentSceneTransitionAction
  | SetCurrentSceneTransitionAction
  | SetSceneTransitionDurationAction
  | GetSceneTransitionCursorAction
  | GetMediaInputStatusAction
  | SetMediaInputCursorAction
  | OffsetMediaInputCursorAction
  | TriggerMediaInputActionAction
  | GetCurrentPreviewSceneAction
  | SetCurrentPreviewSceneAction
  | GetSceneItemLockedAction
  | SetSceneItemLockedAction
  | GetSceneItemIndexAction
  | SetSceneItemIndexAction
  | CreateSceneItemAction
  | RemoveSceneItemAction
  | GetStatsAction
  | GetVersionAction
  | GetHotkeyListAction
  | GetInputPropertiesListPropertyItemsAction
  | PressInputPropertiesButtonAction
  | GetInputAudioBalanceAction
  | SetInputAudioBalanceAction
  | GetInputAudioSyncOffsetAction
  | SetInputAudioSyncOffsetAction
  | GetInputAudioTracksAction
  | SetInputAudioTracksAction
  | DuplicateSceneAction
  | GetSourceScreenshotAction
  | SetCurrentSceneTransitionSettingsAction
  | OpenInputPropertiesDialogAction
  | OpenInputFiltersDialogAction
  | OpenInputInteractDialogAction
  | SetSceneNameAction
  | RemoveSceneAction
  | GetStreamStatusAction
  | StartStreamAction
  | StopStreamAction
  | GetRecordStatusAction
  | StartRecordAction
  | StopRecordAction
  | ToggleRecordPauseAction
  | GetVideoSettingsAction
  | GetSceneItemTransformAction
  | GetSourceFilterAction
  | GetInputVolumeAction
  | GetVirtualCamStatusAction
  | GetReplayBufferStatusAction
  | DuplicateSceneItemAction
  | StopReplayBufferAction
  | GetCurrentProfileAction
  | SetCurrentProfileAction
  | GetCurrentSceneCollectionAction
  | SetCurrentSceneCollectionAction
  | SetStreamInfoAction;